requirements.txt
streamlit
pandas
requests
sqlalchemy
python-dotenv
psycopg2-binary
